backend/pyproject.toml
[project]
name = "chainiq-backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.9.0",
    "pandas>=2.2.0",
    "python-multipart>=0.0.12",
    "psycopg2-binary>=2.9",
]

[project.optional-dependencies]
test = [
    "pytest>=8.1",
    "httpx>=0.27",
    "anyio>=4.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
